refactor(frontend): use error on number inputs
