+++ packages/frontend/src/components/campaign-preview/PreviewBox.vue
